Refactoring: Moved Closing Logic to OnExit
Before: the closing logic for the main window was handled in the Closing_Window method of the SearchViewPage.xaml.cs file. This method was called as an event handler when the main window was closed

Changes Made:
The closing logic, which includes canceling a progress timer and saving application settings, has been moved to the OnExit method in App.xaml.cs. This ensures that the same logic is executed when the application is closed, regardless of whether it's closed by the user or due to an error

This refactoring improves code organization and ensures consistent behavior during application shutdown